requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
python-multipart
pydantic
email-validator
pytest
asyncpg